# Optionally displays vin.prevOut {...} for searchrawtransactions API only.