--- a/btcjson/chainsvrcmds_test.go
+++ b/btcjson/chainsvrcmds_test.go
@@ -679,14 +679,15 @@ func TestChainSvrCmds(t *testing.T) {
 				return btcjson.NewCmd("searchrawtransactions", "1Address")
 			},
 			staticCmd: func() interface{} {
-				return btcjson.NewSearchRawTransactionsCmd("1Address", nil, nil, nil)
+				return btcjson.NewSearchRawTransactionsCmd("1Address", nil, nil, nil, nil)
 			},
 			marshalled: `{"jsonrpc":"1.0","method":"searchrawtransactions","params":["1Address"],"id":1}`,
 			unmarshalled: &btcjson.SearchRawTransactionsCmd{
-				Address: "1Address",
-				Verbose: btcjson.Int(1),
-				Skip:    btcjson.Int(0),
-				Count:   btcjson.Int(100),
+				Address:  "1Address",
+				Verbose:  btcjson.Int(1),
+				Skip:     btcjson.Int(0),
+				Count:    btcjson.Int(100),
+				VinExtra: btcjson.Int(0),
 			},
 		},
 		{
@@ -696,14 +697,15 @@ func TestChainSvrCmds(t *testing.T) {
 			},
 			staticCmd: func() interface{} {
 				return btcjson.NewSearchRawTransactionsCmd("1Address",
-					btcjson.Int(0), nil, nil)
+					btcjson.Int(0), nil, nil, nil)
 			},
 			marshalled: `{"jsonrpc":"1.0","method":"searchrawtransactions","params":["1Address",0],"id":1}`,
 			unmarshalled: &btcjson.SearchRawTransactionsCmd{
-				Address: "1Address",
-				Verbose: btcjson.Int(0),
-				Skip:    btcjson.Int(0),
-				Count:   btcjson.Int(100),
+				Address:  "1Address",
+				Verbose:  btcjson.Int(0),
+				Skip:     btcjson.Int(0),
+				Count:    btcjson.Int(100),
+				VinExtra: btcjson.Int(0),
 			},
 		},
 		{
@@ -713,14 +715,15 @@ func TestChainSvrCmds(t *testing.T) {
 			},
 			staticCmd: func() interface{} {
 				return btcjson.NewSearchRawTransactionsCmd("1Address",
-					btcjson.Int(0), btcjson.Int(5), nil)
+					btcjson.Int(0), btcjson.Int(5), nil, nil)
 			},
 			marshalled: `{"jsonrpc":"1.0","method":"searchrawtransactions","params":["1Address",0,5],"id":1}`,
 			unmarshalled: &btcjson.SearchRawTransactionsCmd{
-				Address: "1Address",
-				Verbose: btcjson.Int(0),
-				Skip:    btcjson.Int(5),
-				Count:   btcjson.Int(100),
+				Address:  "1Address",
+				Verbose:  btcjson.Int(0),
+				Skip:     btcjson.Int(5),
+				Count:    btcjson.Int(100),
+				VinExtra: btcjson.Int(0),
 			},
 		},
 		{
@@ -730,14 +733,33 @@ func TestChainSvrCmds(t *testing.T) {
 			},
 			staticCmd: func() interface{} {
 				return btcjson.NewSearchRawTransactionsCmd("1Address",
-					btcjson.Int(0), btcjson.Int(5), btcjson.Int(10))
+					btcjson.Int(0), btcjson.Int(5), btcjson.Int(10), nil)
 			},
 			marshalled: `{"jsonrpc":"1.0","method":"searchrawtransactions","params":["1Address",0,5,10],"id":1}`,
 			unmarshalled: &btcjson.SearchRawTransactionsCmd{
-				Address: "1Address",
-				Verbose: btcjson.Int(0),
-				Skip:    btcjson.Int(5),
-				Count:   btcjson.Int(10),
+				Address:  "1Address",
+				Verbose:  btcjson.Int(0),
+				Skip:     btcjson.Int(5),
+				Count:    btcjson.Int(10),
+				VinExtra: btcjson.Int(0),
+			},
+		},
+		{
+			name: "searchrawtransactions",
+			newCmd: func() (interface{}, error) {
+				return btcjson.NewCmd("searchrawtransactions", "1Address", 0, 5, 10, 1)
+			},
+			staticCmd: func() interface{} {
+				return btcjson.NewSearchRawTransactionsCmd("1Address",
+					btcjson.Int(0), btcjson.Int(5), btcjson.Int(10), btcjson.Int(1))
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"searchrawtransactions","params":["1Address",0,5,10,1],"id":1}`,
+			unmarshalled: &btcjson.SearchRawTransactionsCmd{
+				Address:  "1Address",
+				Verbose:  btcjson.Int(0),
+				Skip:     btcjson.Int(5),
+				Count:    btcjson.Int(10),
+				VinExtra: btcjson.Int(1),
 			},
 		},
 		{