For background on this pull request, see https://github.com/btcsuite/btcd/issues/485.

## Implementation overview

* searchrawtransactions is modified to return a new type SearchRawTransactionsResult which includes new type VinPrevOut.
* If optional flag vinextra=1 then vin.prevOut struct is returned.  Else vin has same structure as before and no extra processing (to find previous outputs) is performed.
* mempool.fetchInputTransactions is modified to accept a flag includeSpent.  This enables createVinListPrevOut() to retrieve previous output (spent or not) from mempool or db.
* ExtractPkScriptAddrs() is called on the PkScript of the corresponding previous output, if any
* help is modified to display new argument vinextra and new output vin.prevOut.
* test cases were modified to include additional test case using vinextra argument and also test cases for marshalling VinPrevOut.  test coverage in btcjson remains 100%.

## Testing done
I've run a few different addresses against it on both testnet and mainnet, without problems.   Also ran goclean.sh script and all tests pass.

## Commentary
This code is working for my application's use case.  Hopefully it can be incorporated into mainline.

I have one additional searchrawtransctions tweak to add for my use case, which is filtering out inputs/outputs that do not match the input address.  This should result in a large performance gain and reduction in unnecessary data sent over the wire.  I plan to submit that as a separate pull request when ready.

## Usage
<pre>
btcctl help searchrawtransactions
...
Arguments:
...
4. count    (numeric, optional, default=100) The maximum number of transactions to return
5. vinextra (numeric, optional, default=0)   Specify that extra data from previous output will be returned in vin
...
  "vout": n,                   (numeric)         The index of the output being redeemed from the origin transaction (non-coinbase txns only)
  "scriptSig": {               (object)          The signature script used to redeem the origin transaction as a JSON object (non-coinbase txns only)
   "asm": "value",             (string)          Disassembly of the script
   "hex": "value",             (string)          Hex-encoded bytes of the script
  },
  "prevOut": {                 (object)          Data from the origin transaction output with index vout.
   "addresses": ["value",...], (array of string) previous output addresses
   "value": n.nnn,             (numeric)         previous output value
  },
  "sequence": n,               (numeric)         The script sequence number
</pre>

### Example Output, case: flag vinextra = 1
<pre>
$ btcctl searchrawtransactions mpC5xd5UGPmteFbpNfBD2vfhYmb8jWHa1u 1 0 100 1
    "vin": [   
      {
        "txid": "6801cb417573220564c3cec34dd39a0879e24ea75a7ca1ba6a3b8c11c1c6c6b3",
        "vout": 0,
        "scriptSig": {
          "asm": "...",
          "hex": "..."
        },
        "prevOut": {
          "addresses": [
            "mpC5xd5UGPmteFbpNfBD2vfhYmb8jWHa1u"
          ],
          "value": 50
        },
        "sequence": 4294967295
      }
    ],
</pre>

### Example Output, case: flag vinextra = 0   ( or omitted )
<pre>
$ btcctl searchrawtransactions mpC5xd5UGPmteFbpNfBD2vfhYmb8jWHa1u 1 0 100
...
    "vin": [   
      {
        "txid": "6801cb417573220564c3cec34dd39a0879e24ea75a7ca1ba6a3b8c11c1c6c6b3",
        "vout": 0,
        "scriptSig": {
          "asm": "...",
          "hex": "..."
        },
        "sequence": 4294967295
      }
    ],
</pre>