@davecgh 2015-08-20
I've reviewed about half of this so far and want to say very nice work on keeping with the style, adding the appropriate additional tests, etc.

One thing I'd like to request is that the pull request also updates the [searchrawtransactions section of the JSON-RPC API documentation](https://github.com/btcsuite/btcd/blob/master/docs/json_rpc_api.md#searchrawtransactions).

Also, I'm not sure the extra flag to specify whether or not to add the extra prevOut information is needed.  The more important part, in my opinion, was that `getrawtransaction` doesn't attempt to add this information, and, as you know, the pull request uses the separate `SearchRawTransactionsResult` to handle that.

@Roasbeef: Can you review this as well since it's working in this address index area.