@dan-da 2015-08-20
@davecgh hey thanks.  This is my first excursion with both go and btcd, so I'm just trying to go with the flow.

I'll make the changes you requested.  I wasn't aware of the json-rpc docs.  thx for the pointer.

I added the prevout option for a few reasons (1) backwards compatibility.  (2) pushback about the feature. (3) so end-user can avoid the extra processing and bytes if they don't need those values.

In practice it doesn't seem to add much processing overhead when the flag is on, but I would need to test with some large transactions with many non-coinbase inputs to get good perf numbers.

I can remove the flag if that is the consensus.  I haven't seen other API's offering that flag.