@dan-da 2015-08-21
hmm, there seems to be a serious performance problem with my implementation when querying certain addresses with many large transactions.

Without prevOut
<pre>
time btcctl searchrawtransactions 1XPTgDRhN8RFnzniWCddobD9iKZatrvH4 1 0 100000 > /tmp/laz2
real    0m2.721s

$ grep -c blockhash /tmp/laz1
3320
</pre>

With prevOut
<pre>
$time btcctl searchrawtransactions 1XPTgDRhN8RFnzniWCddobD9iKZatrvH4 1 0 100000 > /tmp/laz1
real    8m38.345s

$ grep -c blockhash /tmp/laz2
3320
</pre>

So, ~2 seconds vs ~8 minutes.   Ouch!

Other addresses don't seem to have much of a hit.
<pre>
$ time btcctl searchrawtransactions 1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa 1 0 100000 > /tmp/gen2
real    0m2.334s

$ grep -c blockhash /tmp/gen2
986
</pre>

<pre>
$ time btcctl searchrawtransactions 1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa 1 0 100000 1 > /tmp/gen1
real    0m4.156s

$ grep -c blockhash /tmp/gen1
986
</pre>

So ~2 seconds vs ~4 seconds.   not [as] horrible.

The first address had 3320 transactions and the second one had 986, but the time difference for the prevOut case was enormous:  ~8 minutes vs ~4 secs.  I am guessing that the first had a lot more inputs.

A strange thing is that dstat reports writes of 30-60Mb/sec and 0 reads while the API is running.  The system has free mem and does not appear to be swapping.

I will read up on golang profiling as well as @Roasbeef's implementation recommendations, which I didn't really research previously because I already had a working implementation.

helpful suggestions appreciated....