@davecgh 2015-08-21
Without digging into the specifics a little more I can't say for certain, but I would wager the reason that address is taking so long is because it has a ton of inputs and you're hitting one of the main reasons for PR #380.

To summarize, currently the blocks are stored in leveldb and therefore in order to get a specific transaction means leveldb has to load the entire block since it doesn't provide any way to reach into a specific offset of its values.  That particular address is referencing thousands of transactions which in turn means it's having to load thousands of blocks in order to get at them.

That problem is fixed with #380 since it stores the blocks on the file system instead and therefore allows any piece of a block, whether it be a transaction, header, script, etc, without having to load the entire thing.

That said, while it will certainly be _significantly_ faster with that PR, it's still going to take a little longer to get data for addresses which involve thousands of transactions like the one in question since they will involve thousands of disk seeks to load the relevant transactions.  I've done a significant amount of testing in regards to this particular topic and I'd be willing to bet the performance will be more than acceptable once #380 is in even without any other changes to this PR since I've loaded 10+ thousand transactions across as many blocks with the new code on an old Hitachi 7200RPM disk in a few seconds versus the minutes it takes now to do the same.