@dan-da 2015-08-21
ok, I found the source of slowdown.   as usual PEBKAC.  ;-)

My code was stupidly doing the DB lookup for all transaction inputs for every input, when it only needs to do it once.  After correcting this, the time went from ~8 minutes to ~6 seconds.   And the prevout case is approx 3x slower than without prevout.  Which makes sense because it is doing extra DB work.

without prevOut
<pre>
btcd@btc2:~$ time btcctl searchrawtransactions 1XPTgDRhN8RFnzniWCddobD9iKZatrvH4 1 0 100000  > /tmp/laz-orig
real    0m2.762s
</pre>

with prevOut
<pre>
btcd@btc2:~$ time btcctl searchrawtransactions 1XPTgDRhN8RFnzniWCddobD9iKZatrvH4 1 0 100000 1  > /tmp/laz-prevout-cache
real    0m6.244s
</pre>

This is an address with 3000+ transactions and LOTS of inputs (haven't counted).  So I think that is reasonable.   But the approx 3x slowdown is an argument for keeping the vinextra flag, for folks that don't need the extra data.

I'll update the pull request code in a bit.