@davecgh 2015-08-21
That would certainly do it.  What I said above still applies though, so once that other PR and the requisite supporting code is in, this will benefit from it too.