@dan-da 2015-08-22
ok, I think its ready for review.

Let me know if you want vinextra flag removed.  as stated above, I think there is some value to it, as performance can be worse when loading transactions with lots of inputs.  ( eg ~6.2 seconds vs ~2.7 seconds, as stated above. )

@davecgh I'm looking forward to https://github.com/btcsuite/btcd/pull/380 and friends.  thanks.