@dan-da 2015-08-22
@davecgh , I incorporated your suggested code.  thanks.

btw, I went through the same thought process with int vs bool.  I first implemented it as bool but then changed to int for consistency.