@dan-da 2015-08-22
oh, I tested that the latest code returns correct data for both a coinbase tx and prevOut tx.   Also, tested with flag set to 0,1,2 and omitted.