@davecgh 2015-08-22
Thanks for the updates.  This looks good now code wise to me.  Also, before this can be merged, we'll need to prepare a separate PR for [btcrpcclient](https://github.com/btcsuite/btcrpcclient) which updates it to use the new return type and offer the new flag.

OK pending the separate PR for btcrpcclient.