@dan-da 2015-08-23
requested changes made.   squashed and rebased.   should be good to go...


I don't know anything about btcrpcclient or what changes will be required there.   I definitely would prefer if one of you with experience can do it (probably much faster).  otherwise will look into it when I have some more free time...