@davecgh 2015-08-23
Thanks. I'll make a `btcrpcclient` PR shortly.