@davecgh 2015-08-23
Created btcsuite/btcrpcclient#61 to coincide with this change.